Make ArrayDatum inherit from VarlenDatum

+++ Utils/ChunkIter.cpp
@@ -146,14 +146,14 @@ DEVICE void ChunkIter_get_next(ChunkIter* it,
     if (uncompress && it->type_info.get_compression() != kENCODING_NONE) {
       decompress(it->type_info, it->current_pos, result, &it->datum);
     } else {
-      result->length = it->skip_size;
+      result->length = static_cast<size_t>(it->skip_size);
       result->pointer = it->current_pos;
       result->is_null = it->type_info.is_null(result->pointer);
     }
     it->current_pos += it->skip * it->skip_size;
   } else {
     StringOffsetT offset = *(StringOffsetT*)it->current_pos;
-    result->length = *((StringOffsetT*)it->current_pos + 1) - offset;
+    result->length = static_cast<size_t>(*((StringOffsetT*)it->current_pos + 1) - offset);
     result->pointer = it->second_buf + offset;
     // @TODO(wei) treat zero length as null for now
     result->is_null = (result->length == 0);
@@ -182,14 +182,14 @@ DEVICE void ChunkIter_get_nth(ChunkIter* it,
     if (uncompress && it->type_info.get_compression() != kENCODING_NONE) {
       decompress(it->type_info, current_pos, result, &it->datum);
     } else {
-      result->length = it->skip_size;
+      result->length = static_cast<size_t>(it->skip_size);
       result->pointer = current_pos;
       result->is_null = it->type_info.is_null(result->pointer);
     }
   } else {
     int8_t* current_pos = it->start_pos + n * sizeof(StringOffsetT);
     StringOffsetT offset = *(StringOffsetT*)current_pos;
-    result->length = *((StringOffsetT*)current_pos + 1) - offset;
+    result->length = static_cast<size_t>(*((StringOffsetT*)current_pos + 1) - offset);
     result->pointer = it->second_buf + offset;
     // @TODO(wei) treat zero length as null for now
     result->is_null = (result->length == 0);
@@ -210,13 +210,13 @@ DEVICE void ChunkIter_get_nth(ChunkIter* it, int n, ArrayDatum* result, bool* is
   if (it->skip_size > 0) {
     // for fixed-size
     int8_t* current_pos = it->start_pos + n * it->skip_size;
-    result->length = it->skip_size;
+    result->length = static_cast<size_t>(it->skip_size);
     result->pointer = current_pos;
     result->is_null = it->type_info.is_null(result->pointer);
   } else {
     int8_t* current_pos = it->start_pos + n * sizeof(StringOffsetT);
     StringOffsetT offset = *(StringOffsetT*)current_pos;
-    result->length = *((StringOffsetT*)current_pos + 1) - offset;
+    result->length = static_cast<size_t>(*((StringOffsetT*)current_pos + 1) - offset);
     result->pointer = it->second_buf + offset;
     // @TODO(wei) treat zero length as null for now
     result->is_null = (result->length == 0);
